+++ src/layout/section/highlightSection/HighlightSection.tsx
@@ -1,15 +1,7 @@
-import React, { Fragment } from 'react'
+import React, { Fragment, useState } from 'react'
 
-import { Toggle, useToggleContext } from './Accordion'
-import {
-  Col,
-  Grid,
-  HighlightSectionContent,
-  HighlightSectionItem,
-  HighlightSectionLink,
-  HighlightSectionTitle,
-  StyledHighlightSection,
-} from './HighlightSection.style'
+import { Col, Grid, HighlightSectionElements } from './HighlightSection.style'
+import { Toggle } from './Toggle'
 
 export type HighlightSectionProps = Readonly<{
   className?: string
@@ -22,9 +14,9 @@ export type HighlightSectionProps = Readonly<{
  * A specialized section with an highlighting background color.
  */
 export const GridListItems = ({ items }) => {
-  const listItems = items.map(({ id, label, data, href, ariaLabel }) => (
-    <Col key={id}>
-      <HighlightSectionItem label={label} data={data} href={href} ariaLabel={ariaLabel} />
+  const listItems = items.map(({ id, label, data, hidden, href, ariaLabel }) => (
+    <Col key={id} hidden={hidden}>
+      <HighlightSectionElements.Item label={label} data={data} href={href} ariaLabel={ariaLabel} />
     </Col>
   ))
   return <Grid>{listItems}</Grid>
@@ -32,10 +24,11 @@ export const GridListItems = ({ items }) => {
 
 export const HighlightContentItems = ({ heading, items, ...props }) => (
   <Fragment>
-    {heading && <HighlightSectionTitle as="h2">{heading}</HighlightSectionTitle>}
+    {heading && <HighlightSectionElements.Title as="h2">{heading}</HighlightSectionElements.Title>}
     <GridListItems items={items} />
   </Fragment>
 )
+
 export const HighlightSection = ({
   content,
   toggleLabel,
@@ -43,38 +36,36 @@ export const HighlightSection = ({
   children,
 }: HighlightSectionProps) => {
   const DEFAULT_ITEMS_SIZE = 3
-  const handleStartItems = (startItems: Array<Object>) => {
-    const { on } = useToggleContext()
-    console.log('on', on)
-    const defaultItems = startItems.filter((item, index) =>
-      index < DEFAULT_ITEMS_SIZE ? item : null,
-    )
-    return on ? defaultItems : startItems
-  }
+  const [isToggled, setOnToggle] = useState(false)
+  const startItems = content[0].items
+
+  const defaultItems = startItems.filter((item, index) => {
+    if (index < DEFAULT_ITEMS_SIZE) {
+      item.hidden = !isToggled
+    }
+    return item
+  })
+
+  const items = isToggled ? defaultItems : startItems
 
   return (
-    <StyledHighlightSection className={className}>
-      <HighlightSectionContent>
-        <Toggle onToggle={on => handleStartItems(content[0].items)}>
-          <HighlightContentItems
-            key="bus"
-            heading={content[0].heading}
-            items={handleStartItems(content[0].items)}
-          />
-          <Toggle.On>
+    <HighlightSectionElements.Section className={className}>
+      <HighlightSectionElements.Content>
+        <Toggle onToggle={on => setOnToggle(on)}>
+          <HighlightContentItems key="bus" heading={content[0].heading} items={items} />
+          <Toggle.On aria-expanded={isToggled}>
             <HighlightContentItems
               key="trajets"
               heading={content[1].heading}
               items={content[1].items}
             />
           </Toggle.On>
-          <Toggle.Button element={HighlightSectionLink} ariaLabel>
+          <Toggle.Button element={HighlightSectionElements.Link} role="button">
             <Toggle.Off>{toggleLabel.off}</Toggle.Off>
             <Toggle.On>{toggleLabel.on}</Toggle.On>
           </Toggle.Button>
         </Toggle>
-        {children}
-      </HighlightSectionContent>
-    </StyledHighlightSection>
+      </HighlightSectionElements.Content>
+    </HighlightSectionElements.Section>
   )
 }