+++ src/layout/section/highlightSection/Toggle.tsx
@@ -1,7 +1,10 @@
 import React, { useContext, useMemo } from 'react'
+import styled from 'styled-components'
 
-// TOGGLE SOLUTION
 const ToggleContext = React.createContext()
+const Hidden = styled('div')`
+  display: none;
+`
 
 function useEffectAfterMount(cb, dependencies) {
   const justMounted = React.useRef(true)
@@ -21,28 +24,29 @@ const Toggle = props => {
   useEffectAfterMount(() => {
     props.onToggle(on)
   }, [on])
+
   const value = React.useMemo(() => ({ on, toggle }), [on])
   return <ToggleContext.Provider value={value}>{props.children}</ToggleContext.Provider>
 }
 
 export const useToggleContext = () => {
   const context = React.useContext(ToggleContext)
-  // if (!context) {
-  //   throw new Error('Toggle compound components cannot be rendered outside the Toggle component')
-  // }
+  if (!context) {
+    throw new Error('Toggle compound components cannot be rendered outside the Toggle component')
+  }
   return context
 }
 
 function On({ children }) {
   const { on } = useToggleContext()
 
-  return on ? children : null
+  return on ? children : <Hidden>{children}</Hidden>
 }
 
 function Off({ children }) {
   const { on } = useToggleContext()
 
-  return on ? null : children
+  return on ? <Hidden>{children}</Hidden> : children
 }
 
 function Button({ element: Component, children, ...otherProps }) {