+++ src/layout/section/highlightSection/index.tsx
@@ -1,2 +1,2 @@
-export { HighlightSection, HighlightContentItems, HighlightSectionProps } from './HighlightSection'
-export { HighlightSectionLink } from './HighlightSection.style'
+export { HighlightSection, HighlightSectionProps } from './HighlightSection'
+export { HighlightSectionElements } from './HighlightSection.style'